// clients/js/src/generated/accounts/eventPool.ts
export type EventPool = {
  discriminator: ReadonlyUint8Array;
  version: number;
  coachAuthority: Address;
  payoutRecipient: Address;
  paymentMint: Address;
  vault: Address;
  nonce: bigint;
  priceEurcBaseUnits: bigint;
  minimumParticipants: number;
  maximumParticipants: number;
  participantCount: number;
  totalFundedBaseUnits: bigint;
  totalRefundedBaseUnits: bigint;
  fundingDeadline: bigint;
  eventStartAt: bigint;
  eventEndAt: bigint;
  createdAt: bigint;
  status: EventPoolStatus;
  settledAt: Option<bigint>;
  paidAt: Option<bigint>;
  bump: number;
  vaultBump: number;
  reserved: Array<number>;
};