// clients/js/src/generated/types/eventPoolStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum EventPoolStatus {
  Funding,
  Succeeded,
  Paid,
  Failed,
}

export type EventPoolStatusArgs = EventPoolStatus;

export function getEventPoolStatusEncoder(): FixedSizeEncoder<EventPoolStatusArgs> {
  return getEnumEncoder(EventPoolStatus);
}

export function getEventPoolStatusDecoder(): FixedSizeDecoder<EventPoolStatus> {
  return getEnumDecoder(EventPoolStatus);
}

export function getEventPoolStatusCodec(): FixedSizeCodec<
  EventPoolStatusArgs,
  EventPoolStatus
> {
  return combineCodec(getEventPoolStatusEncoder(), getEventPoolStatusDecoder());
}